pyproject.toml
[project]
name = "pdf-reader-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"


dependencies = [
    "google-genai>=1.15.0",
    "google-generativeai>=0.8.5",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.6",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.9.1",
    "ruff>=0.7.0",
    "mypy>=1.16.0",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = ["E", "F", "I", "ARG", "B", "C4", "SIM", "UP", "RUF", "N"]
